fix UT (#64616)
