+++ be/test/format_v2/expr/cast_test.cpp
@@ -34,31 +34,11 @@
 #include "exprs/vexpr_context.h"
 #include "exprs/vliteral.h"
 #include "exprs/vslot_ref.h"
-#include "format_v2/column_mapper.h"
-#include "format_v2/file_reader.h"
-#include "format_v2/table_reader.h"
 #include "runtime/descriptors.h"
 #include "testutil/column_helper.h"
-#include "testutil/mock/mock_in_expr.h"
 #include "testutil/mock/mock_runtime_state.h"
 
 namespace doris::format {
-namespace {
-
-format::LocalColumnIndex field_projection(int32_t column_id) {
-    return format::LocalColumnIndex {.index = column_id};
-}
-
-std::vector<int32_t> projection_ids(const std::vector<format::LocalColumnIndex>& projections) {
-    std::vector<int32_t> ids;
-    ids.reserve(projections.size());
-    for (const auto& projection : projections) {
-        ids.push_back(projection.index);
-    }
-    return ids;
-}
-
-} // namespace
 
 class CastTest : public testing::Test {
 protected:
@@ -79,77 +59,6 @@ protected:
     }
 
     MockRuntimeState state;
-};
-
-class Int64ChildGreaterThanExpr final : public VExpr {
-public:
-    explicit Int64ChildGreaterThanExpr(int64_t value)
-            : VExpr(std::make_shared<DataTypeUInt8>(), false), _value(value) {}
-
-    Status execute_column_impl(VExprContext* context, const Block* block, const Selector* selector,
-                               size_t count, ColumnPtr& result_column) const override {
-        ColumnPtr child_column;
-        RETURN_IF_ERROR(
-                get_child(0)->execute_column(context, block, selector, count, child_column));
-        const auto& input = assert_cast<const ColumnInt64&>(*child_column);
-        auto result = ColumnUInt8::create();
-        auto& result_data = result->get_data();
-        result_data.resize(count);
-        for (size_t row = 0; row < count; ++row) {
-            result_data[row] = input.get_element(row) > _value;
-        }
-        result_column = std::move(result);
-        return Status::OK();
-    }
-
-    const std::string& expr_name() const override { return _expr_name; }
-
-private:
-    const int64_t _value;
-    const std::string _expr_name = "Int64ChildGreaterThanExpr";
-};
-
-class Int64BinaryPredicateExpr final : public VExpr {
-public:
-    explicit Int64BinaryPredicateExpr(TExprOpcode::type opcode)
-            : VExpr(std::make_shared<DataTypeUInt8>(), false) {
-        set_node_type(TExprNodeType::BINARY_PRED);
-        _opcode = opcode;
-    }
-
-    Status execute_column_impl(VExprContext* context, const Block* block, const Selector* selector,
-                               size_t count, ColumnPtr& result_column) const override {
-        ColumnPtr left_column;
-        RETURN_IF_ERROR(get_child(0)->execute_column(context, block, selector, count, left_column));
-        ColumnPtr right_column;
-        RETURN_IF_ERROR(
-                get_child(1)->execute_column(context, block, selector, count, right_column));
-
-        auto result = ColumnUInt8::create();
-        auto& result_data = result->get_data();
-        result_data.resize(count);
-        for (size_t row = 0; row < count; ++row) {
-            const auto left = left_column->get_int(row);
-            const auto right = right_column->get_int(row);
-            switch (_opcode) {
-            case TExprOpcode::GT:
-                result_data[row] = left > right;
-                break;
-            case TExprOpcode::LT:
-                result_data[row] = left < right;
-                break;
-            default:
-                return Status::InternalError("Unsupported test opcode {}", _opcode);
-            }
-        }
-        result_column = std::move(result);
-        return Status::OK();
-    }
-
-    const std::string& expr_name() const override { return _expr_name; }
-
-private:
-    const std::string _expr_name = "Int64BinaryPredicateExpr";
 };
 
 TEST_F(CastTest, CastIntSlotToBigInt) {
@@ -258,748 +167,6 @@ TEST_F(CastTest, PrepareRejectsMultipleChildren) {
     auto status = context.prepare(&state, RowDescriptor());
     ASSERT_FALSE(status.ok());
     EXPECT_NE(status.to_string().find("exactly 1 child expr"), std::string::npos);
-}
-
-TEST_F(CastTest, ColumnMapperBuildsCastProjectionForTypeMismatch) {
-    format::TableColumnMapper mapper;
-    format::ColumnDefinition table_column;
-    table_column.identifier = Field::create_field<TYPE_INT>(7);
-    table_column.name = "value";
-    table_column.type = std::make_shared<DataTypeInt64>();
-    std::vector<format::ColumnDefinition> projected_columns {table_column};
-
-    format::ColumnDefinition file_field;
-    file_field.identifier = Field::create_field<TYPE_INT>(0);
-    file_field.name = "value";
-    file_field.type = std::make_shared<DataTypeInt32>();
-    std::vector<format::ColumnDefinition> file_schema {file_field};
-
-    auto status = mapper.create_mapping(projected_columns, {}, file_schema);
-    ASSERT_TRUE(status.ok()) << status;
-    ASSERT_EQ(mapper.mappings().size(), 1);
-    format::FileScanRequest file_request;
-    status = mapper.create_scan_request({}, {}, projected_columns, &file_request);
-    ASSERT_TRUE(status.ok()) << status;
-    const auto& mapping = mapper.mappings()[0];
-    EXPECT_FALSE(mapping.is_trivial);
-    ASSERT_NE(mapping.projection, nullptr);
-
-    Block block;
-    block.insert(ColumnHelper::create_column_with_name<DataTypeInt32>({11, 22}));
-    int result_column_id = -1;
-    status = prepare_open_execute(mapping.projection.get(), &block, &result_column_id);
-    ASSERT_TRUE(status.ok()) << status;
-
-    const auto& result_column =
-            assert_cast<const ColumnInt64&>(*block.get_by_position(result_column_id).column);
-    EXPECT_EQ(result_column.get_data()[0], 11);
-    EXPECT_EQ(result_column.get_data()[1], 22);
-
-    mapping.projection->close();
-}
-
-TEST_F(CastTest, ColumnMapperTreatsEquivalentTypesAsTrivial) {
-    format::TableColumnMapper mapper;
-    format::ColumnDefinition table_column;
-    table_column.identifier = Field::create_field<TYPE_INT>(7);
-    table_column.name = "value";
-    table_column.type = std::make_shared<DataTypeInt32>();
-    std::vector<format::ColumnDefinition> projected_columns {table_column};
-
-    format::ColumnDefinition file_field;
-    file_field.identifier = Field::create_field<TYPE_INT>(0);
-    file_field.name = "value";
-    file_field.type = std::make_shared<DataTypeInt32>();
-    std::vector<format::ColumnDefinition> file_schema {file_field};
-
-    auto status = mapper.create_mapping(projected_columns, {}, file_schema);
-    ASSERT_TRUE(status.ok()) << status;
-    ASSERT_EQ(mapper.mappings().size(), 1);
-    EXPECT_TRUE(mapper.mappings()[0].is_trivial);
-}
-
-TEST_F(CastTest, ColumnMapperBuildsCastFilterForTypeMismatch) {
-    format::TableColumnMapper mapper;
-    format::ColumnDefinition table_column;
-    table_column.identifier = Field::create_field<TYPE_INT>(7);
-    table_column.name = "value";
-    table_column.type = std::make_shared<DataTypeInt64>();
-    std::vector<format::ColumnDefinition> projected_columns {table_column};
-
-    format::ColumnDefinition file_field;
-    file_field.identifier = Field::create_field<TYPE_INT>(0);
-    file_field.name = "value";
-    file_field.type = std::make_shared<DataTypeInt32>();
-    std::vector<format::ColumnDefinition> file_schema {file_field};
-
-    auto status = mapper.create_mapping(projected_columns, {}, file_schema);
-    ASSERT_TRUE(status.ok()) << status;
-
-    auto predicate = std::make_shared<Int64ChildGreaterThanExpr>(15);
-    predicate->add_child(VSlotRef::create_shared(0, 0, -1, table_column.type, "value"));
-    format::TableFilter table_filter;
-    table_filter.conjunct = VExprContext::create_shared(predicate);
-    table_filter.global_indices = {format::GlobalIndex(0)};
-
-    format::FileScanRequest file_request;
-    ASSERT_TRUE(
-            mapper.create_scan_request({table_filter}, {}, projected_columns, &file_request, &state)
-                    .ok());
-    ASSERT_EQ(file_request.conjuncts.size(), 1);
-    ASSERT_EQ(projection_ids(file_request.predicate_columns), std::vector<int32_t>({0}));
-    const auto& localized_expr = file_request.conjuncts[0]->root();
-    ASSERT_EQ(localized_expr->get_num_children(), 1);
-    const auto& localized_child = localized_expr->children()[0];
-    ASSERT_NE(dynamic_cast<const Cast*>(localized_child.get()), nullptr);
-    ASSERT_EQ(localized_child->get_num_children(), 1);
-    const auto* localized_slot = assert_cast<const VSlotRef*>(localized_child->children()[0].get());
-    EXPECT_EQ(localized_slot->column_id(), 0);
-    EXPECT_TRUE(localized_slot->data_type()->equals(*file_field.type));
-    EXPECT_TRUE(localized_child->data_type()->equals(*table_column.type));
-
-    Block block;
-    block.insert(ColumnHelper::create_column_with_name<DataTypeInt32>({11, 22}));
-    auto* conjunct = file_request.conjuncts[0].get();
-    status = conjunct->prepare(&state, RowDescriptor());
-    ASSERT_TRUE(status.ok()) << status;
-    status = conjunct->open(&state);
-    ASSERT_TRUE(status.ok()) << status;
-    IColumn::Filter filter(block.rows(), 1);
-    bool can_filter_all = false;
-    status = conjunct->execute_filter(&block, filter.data(), block.rows(), false, &can_filter_all);
-    ASSERT_TRUE(status.ok()) << status;
-    EXPECT_FALSE(can_filter_all);
-    ASSERT_EQ(filter.size(), 2);
-    EXPECT_EQ(filter[0], 0);
-    EXPECT_EQ(filter[1], 1);
-
-    file_request.conjuncts[0]->close();
-}
-
-TEST_F(CastTest, ColumnMapperRepreparesRewrittenPreparedFilter) {
-    format::TableColumnMapper mapper;
-    format::ColumnDefinition table_column;
-    table_column.identifier = Field::create_field<TYPE_INT>(7);
-    table_column.name = "value";
-    table_column.type = std::make_shared<DataTypeInt64>();
-    std::vector<format::ColumnDefinition> projected_columns {table_column};
-
-    format::ColumnDefinition file_field;
-    file_field.identifier = Field::create_field<TYPE_INT>(0);
-    file_field.name = "value";
-    file_field.type = std::make_shared<DataTypeInt32>();
-    std::vector<format::ColumnDefinition> file_schema {file_field};
-
-    auto status = mapper.create_mapping(projected_columns, {}, file_schema);
-    ASSERT_TRUE(status.ok()) << status;
-
-    auto cast = Cast::create_shared(table_column.type);
-    cast->add_child(VSlotRef::create_shared(0, 0, -1, table_column.type, "value"));
-    format::TableFilter table_filter;
-    table_filter.conjunct = VExprContext::create_shared(cast);
-    table_filter.global_indices = {format::GlobalIndex(0)};
-    status = table_filter.conjunct->prepare(&state, RowDescriptor());
-    ASSERT_TRUE(status.ok()) << status;
-    status = table_filter.conjunct->open(&state);
-    ASSERT_TRUE(status.ok()) << status;
-
-    format::FileScanRequest file_request;
-    ASSERT_TRUE(
-            mapper.create_scan_request({table_filter}, {}, projected_columns, &file_request, &state)
-                    .ok());
-    ASSERT_EQ(file_request.conjuncts.size(), 1);
-    const auto& localized_expr = file_request.conjuncts[0]->root();
-    ASSERT_NE(dynamic_cast<const Cast*>(localized_expr.get()), nullptr);
-    ASSERT_EQ(localized_expr->get_num_children(), 1);
-    const auto* localized_slot = assert_cast<const VSlotRef*>(localized_expr->children()[0].get());
-    EXPECT_EQ(localized_slot->column_id(), 0);
-    EXPECT_TRUE(localized_slot->data_type()->equals(*file_field.type));
-
-    status = file_request.conjuncts[0]->prepare(&state, RowDescriptor());
-    ASSERT_TRUE(status.ok()) << status;
-    status = file_request.conjuncts[0]->open(&state);
-    ASSERT_TRUE(status.ok()) << status;
-
-    file_request.conjuncts[0]->close();
-}
-
-TEST_F(CastTest, ColumnMapperCastsLiteralForSlotLiteralPredicateTypeMismatch) {
-    format::TableColumnMapper mapper;
-    format::ColumnDefinition table_column;
-    table_column.identifier = Field::create_field<TYPE_INT>(7);
-    table_column.name = "value";
-    table_column.type = std::make_shared<DataTypeInt64>();
-    std::vector<format::ColumnDefinition> projected_columns {table_column};
-
-    format::ColumnDefinition file_field;
-    file_field.identifier = Field::create_field<TYPE_INT>(0);
-    file_field.name = "value";
-    file_field.type = std::make_shared<DataTypeInt32>();
-    std::vector<format::ColumnDefinition> file_schema {file_field};
-
-    auto status = mapper.create_mapping(projected_columns, {}, file_schema);
-    ASSERT_TRUE(status.ok()) << status;
-
-    auto predicate = std::make_shared<Int64BinaryPredicateExpr>(TExprOpcode::GT);
-    predicate->add_child(VSlotRef::create_shared(0, 0, -1, table_column.type, "value"));
-    predicate->add_child(
-            VLiteral::create_shared(table_column.type, Field::create_field<TYPE_BIGINT>(15)));
-    format::TableFilter table_filter;
-    table_filter.conjunct = VExprContext::create_shared(predicate);
-    table_filter.global_indices = {format::GlobalIndex(0)};
-
-    format::FileScanRequest file_request;
-    ASSERT_TRUE(
-            mapper.create_scan_request({table_filter}, {}, projected_columns, &file_request, &state)
-                    .ok());
-    ASSERT_EQ(file_request.conjuncts.size(), 1);
-    ASSERT_EQ(projection_ids(file_request.predicate_columns), std::vector<int32_t>({0}));
-    const auto& localized_expr = file_request.conjuncts[0]->root();
-    ASSERT_EQ(localized_expr->get_num_children(), 2);
-    const auto* localized_slot = assert_cast<const VSlotRef*>(localized_expr->children()[0].get());
-    EXPECT_EQ(localized_slot->column_id(), 0);
-    EXPECT_TRUE(localized_slot->data_type()->equals(*file_field.type));
-    const auto& localized_literal = localized_expr->children()[1];
-    EXPECT_TRUE(localized_literal->is_literal());
-    EXPECT_TRUE(localized_literal->data_type()->equals(*file_field.type));
-
-    Block block;
-    block.insert(ColumnHelper::create_column_with_name<DataTypeInt32>({11, 22}));
-    auto* conjunct = file_request.conjuncts[0].get();
-    status = conjunct->prepare(&state, RowDescriptor());
-    ASSERT_TRUE(status.ok()) << status;
-    status = conjunct->open(&state);
-    ASSERT_TRUE(status.ok()) << status;
-    IColumn::Filter filter(block.rows(), 1);
-    bool can_filter_all = false;
-    status = conjunct->execute_filter(&block, filter.data(), block.rows(), false, &can_filter_all);
-    ASSERT_TRUE(status.ok()) << status;
-    EXPECT_FALSE(can_filter_all);
-    ASSERT_EQ(filter.size(), 2);
-    EXPECT_EQ(filter[0], 0);
-    EXPECT_EQ(filter[1], 1);
-
-    file_request.conjuncts[0]->close();
-}
-
-TEST_F(CastTest, ColumnMapperCastsLiteralForLiteralSlotPredicateTypeMismatch) {
-    format::TableColumnMapper mapper;
-    format::ColumnDefinition table_column;
-    table_column.identifier = Field::create_field<TYPE_INT>(7);
-    table_column.name = "value";
-    table_column.type = std::make_shared<DataTypeInt64>();
-    std::vector<format::ColumnDefinition> projected_columns {table_column};
-
-    format::ColumnDefinition file_field;
-    file_field.identifier = Field::create_field<TYPE_INT>(0);
-    file_field.name = "value";
-    file_field.type = std::make_shared<DataTypeInt32>();
-    std::vector<format::ColumnDefinition> file_schema {file_field};
-
-    auto status = mapper.create_mapping(projected_columns, {}, file_schema);
-    ASSERT_TRUE(status.ok()) << status;
-
-    auto predicate = std::make_shared<Int64BinaryPredicateExpr>(TExprOpcode::LT);
-    predicate->add_child(
-            VLiteral::create_shared(table_column.type, Field::create_field<TYPE_BIGINT>(15)));
-    predicate->add_child(VSlotRef::create_shared(0, 0, -1, table_column.type, "value"));
-    format::TableFilter table_filter;
-    table_filter.conjunct = VExprContext::create_shared(predicate);
-    table_filter.global_indices = {format::GlobalIndex(0)};
-
-    format::FileScanRequest file_request;
-    ASSERT_TRUE(
-            mapper.create_scan_request({table_filter}, {}, projected_columns, &file_request, &state)
-                    .ok());
-    ASSERT_EQ(file_request.conjuncts.size(), 1);
-    const auto& localized_expr = file_request.conjuncts[0]->root();
-    ASSERT_EQ(localized_expr->get_num_children(), 2);
-    const auto& localized_literal = localized_expr->children()[0];
-    EXPECT_TRUE(localized_literal->is_literal());
-    EXPECT_TRUE(localized_literal->data_type()->equals(*file_field.type));
-    const auto* localized_slot = assert_cast<const VSlotRef*>(localized_expr->children()[1].get());
-    EXPECT_EQ(localized_slot->column_id(), 0);
-    EXPECT_TRUE(localized_slot->data_type()->equals(*file_field.type));
-
-    Block block;
-    block.insert(ColumnHelper::create_column_with_name<DataTypeInt32>({11, 22}));
-    auto* conjunct = file_request.conjuncts[0].get();
-    status = conjunct->prepare(&state, RowDescriptor());
-    ASSERT_TRUE(status.ok()) << status;
-    status = conjunct->open(&state);
-    ASSERT_TRUE(status.ok()) << status;
-    IColumn::Filter filter(block.rows(), 1);
-    bool can_filter_all = false;
-    status = conjunct->execute_filter(&block, filter.data(), block.rows(), false, &can_filter_all);
-    ASSERT_TRUE(status.ok()) << status;
-    EXPECT_FALSE(can_filter_all);
-    ASSERT_EQ(filter.size(), 2);
-    EXPECT_EQ(filter[0], 0);
-    EXPECT_EQ(filter[1], 1);
-
-    file_request.conjuncts[0]->close();
-}
-
-TEST_F(CastTest, ColumnMapperCastsInPredicateLiteralsForTypeMismatch) {
-    format::TableColumnMapper mapper;
-    format::ColumnDefinition table_column;
-    table_column.identifier = Field::create_field<TYPE_INT>(7);
-    table_column.name = "value";
-    table_column.type = std::make_shared<DataTypeInt64>();
-    std::vector<format::ColumnDefinition> projected_columns {table_column};
-
-    format::ColumnDefinition file_field;
-    file_field.identifier = Field::create_field<TYPE_INT>(0);
-    file_field.name = "value";
-    file_field.type = std::make_shared<DataTypeInt32>();
-    std::vector<format::ColumnDefinition> file_schema {file_field};
-
-    auto status = mapper.create_mapping(projected_columns, {}, file_schema);
-    ASSERT_TRUE(status.ok()) << status;
-
-    auto predicate = MockInExpr::create();
-    predicate->add_child(VSlotRef::create_shared(0, 0, -1, table_column.type, "value"));
-    predicate->add_child(
-            VLiteral::create_shared(table_column.type, Field::create_field<TYPE_BIGINT>(15)));
-    predicate->add_child(
-            VLiteral::create_shared(table_column.type, Field::create_field<TYPE_BIGINT>(22)));
-    format::TableFilter table_filter;
-    table_filter.conjunct = VExprContext::create_shared(predicate);
-    table_filter.global_indices = {format::GlobalIndex(0)};
-
-    format::FileScanRequest file_request;
-    ASSERT_TRUE(
-            mapper.create_scan_request({table_filter}, {}, projected_columns, &file_request, &state)
-                    .ok());
-    ASSERT_EQ(file_request.conjuncts.size(), 1);
-    ASSERT_EQ(projection_ids(file_request.predicate_columns), std::vector<int32_t>({0}));
-    const auto& localized_expr = file_request.conjuncts[0]->root();
-    ASSERT_EQ(localized_expr->get_num_children(), 3);
-    const auto* localized_slot = assert_cast<const VSlotRef*>(localized_expr->children()[0].get());
-    EXPECT_EQ(localized_slot->column_id(), 0);
-    EXPECT_TRUE(localized_slot->data_type()->equals(*file_field.type));
-    EXPECT_TRUE(localized_expr->children()[1]->is_literal());
-    EXPECT_TRUE(localized_expr->children()[1]->data_type()->equals(*file_field.type));
-    EXPECT_TRUE(localized_expr->children()[2]->is_literal());
-    EXPECT_TRUE(localized_expr->children()[2]->data_type()->equals(*file_field.type));
-}
-
-TEST_F(CastTest, ColumnMapperFallsBackToSlotCastWhenInPredicateLiteralRewriteFails) {
-    format::TableColumnMapper mapper;
-    format::ColumnDefinition table_column;
-    table_column.identifier = Field::create_field<TYPE_INT>(7);
-    table_column.name = "value";
-    table_column.type = std::make_shared<DataTypeString>();
-    std::vector<format::ColumnDefinition> projected_columns {table_column};
-
-    format::ColumnDefinition file_field;
-    file_field.identifier = Field::create_field<TYPE_INT>(0);
-    file_field.name = "value";
-    file_field.type = std::make_shared<DataTypeInt32>();
-    std::vector<format::ColumnDefinition> file_schema {file_field};
-
-    auto status = mapper.create_mapping(projected_columns, {}, file_schema);
-    ASSERT_TRUE(status.ok()) << status;
-
-    auto predicate = MockInExpr::create();
-    predicate->add_child(VSlotRef::create_shared(0, 0, -1, table_column.type, "value"));
-    predicate->add_child(
-            VLiteral::create_shared(table_column.type, Field::create_field<TYPE_STRING>("10")));
-    predicate->add_child(
-            VLiteral::create_shared(table_column.type, Field::create_field<TYPE_STRING>("bad")));
-    format::TableFilter table_filter;
-    table_filter.conjunct = VExprContext::create_shared(predicate);
-    table_filter.global_indices = {format::GlobalIndex(0)};
-
-    format::FileScanRequest file_request;
-    ASSERT_TRUE(
-            mapper.create_scan_request({table_filter}, {}, projected_columns, &file_request, &state)
-                    .ok());
-    ASSERT_EQ(file_request.conjuncts.size(), 1);
-    const auto& localized_expr = file_request.conjuncts[0]->root();
-    ASSERT_EQ(localized_expr->get_num_children(), 3);
-    const auto& localized_child = localized_expr->children()[0];
-    ASSERT_NE(dynamic_cast<const Cast*>(localized_child.get()), nullptr);
-    ASSERT_EQ(localized_child->get_num_children(), 1);
-    const auto* localized_slot = assert_cast<const VSlotRef*>(localized_child->children()[0].get());
-    EXPECT_EQ(localized_slot->column_id(), 0);
-    EXPECT_TRUE(localized_slot->data_type()->equals(*file_field.type));
-    EXPECT_TRUE(localized_child->data_type()->equals(*table_column.type));
-    EXPECT_TRUE(localized_expr->children()[1]->is_literal());
-    EXPECT_TRUE(localized_expr->children()[1]->data_type()->equals(*table_column.type));
-    EXPECT_TRUE(localized_expr->children()[2]->is_literal());
-    EXPECT_TRUE(localized_expr->children()[2]->data_type()->equals(*table_column.type));
-}
-
-TEST_F(CastTest, ColumnMapperDoesNotLeakRewrittenInPredicateLiteralAcrossSplits) {
-    format::ColumnDefinition table_column;
-    table_column.identifier = Field::create_field<TYPE_INT>(7);
-    table_column.name = "value";
-    table_column.type = std::make_shared<DataTypeInt64>();
-    std::vector<format::ColumnDefinition> projected_columns {table_column};
-
-    auto predicate = MockInExpr::create();
-    predicate->add_child(VSlotRef::create_shared(0, 0, -1, table_column.type, "value"));
-    predicate->add_child(
-            VLiteral::create_shared(table_column.type, Field::create_field<TYPE_BIGINT>(15)));
-    predicate->add_child(
-            VLiteral::create_shared(table_column.type, Field::create_field<TYPE_BIGINT>(22)));
-    format::TableFilter table_filter;
-    table_filter.conjunct = VExprContext::create_shared(predicate);
-    table_filter.global_indices = {format::GlobalIndex(0)};
-
-    format::ColumnDefinition int_file_field;
-    int_file_field.identifier = Field::create_field<TYPE_INT>(0);
-    int_file_field.name = "value";
-    int_file_field.type = std::make_shared<DataTypeInt32>();
-    format::TableColumnMapper int_mapper;
-    ASSERT_TRUE(int_mapper.create_mapping(projected_columns, {}, {int_file_field}).ok());
-    format::FileScanRequest int_request;
-    ASSERT_TRUE(int_mapper
-                        .create_scan_request({table_filter}, {}, projected_columns, &int_request,
-                                             &state)
-                        .ok());
-    ASSERT_EQ(int_request.conjuncts.size(), 1);
-    const auto& int_localized_expr = int_request.conjuncts[0]->root();
-    ASSERT_EQ(int_localized_expr->get_num_children(), 3);
-    EXPECT_TRUE(int_localized_expr->children()[1]->is_literal());
-    EXPECT_TRUE(int_localized_expr->children()[1]->data_type()->equals(*int_file_field.type));
-    EXPECT_TRUE(int_localized_expr->children()[2]->is_literal());
-    EXPECT_TRUE(int_localized_expr->children()[2]->data_type()->equals(*int_file_field.type));
-
-    format::ColumnDefinition bigint_file_field;
-    bigint_file_field.identifier = Field::create_field<TYPE_INT>(0);
-    bigint_file_field.name = "value";
-    bigint_file_field.type = std::make_shared<DataTypeInt64>();
-    format::TableColumnMapper bigint_mapper;
-    ASSERT_TRUE(bigint_mapper.create_mapping(projected_columns, {}, {bigint_file_field}).ok());
-    format::FileScanRequest bigint_request;
-    ASSERT_TRUE(bigint_mapper
-                        .create_scan_request({table_filter}, {}, projected_columns, &bigint_request,
-                                             &state)
-                        .ok());
-    ASSERT_EQ(bigint_request.conjuncts.size(), 1);
-    const auto& bigint_localized_expr = bigint_request.conjuncts[0]->root();
-    ASSERT_EQ(bigint_localized_expr->get_num_children(), 3);
-    const auto* localized_slot =
-            assert_cast<const VSlotRef*>(bigint_localized_expr->children()[0].get());
-    EXPECT_EQ(localized_slot->column_id(), 0);
-    EXPECT_TRUE(localized_slot->data_type()->equals(*bigint_file_field.type));
-    EXPECT_TRUE(bigint_localized_expr->children()[1]->is_literal());
-    EXPECT_TRUE(bigint_localized_expr->children()[1]->data_type()->equals(*bigint_file_field.type));
-    EXPECT_TRUE(bigint_localized_expr->children()[2]->is_literal());
-    EXPECT_TRUE(bigint_localized_expr->children()[2]->data_type()->equals(*bigint_file_field.type));
-}
-
-TEST_F(CastTest, ColumnMapperFallsBackToSlotCastWhenLiteralRewriteFails) {
-    format::TableColumnMapper mapper;
-    format::ColumnDefinition table_column;
-    table_column.identifier = Field::create_field<TYPE_INT>(7);
-    table_column.name = "value";
-    table_column.type = std::make_shared<DataTypeString>();
-    std::vector<format::ColumnDefinition> projected_columns {table_column};
-
-    format::ColumnDefinition file_field;
-    file_field.identifier = Field::create_field<TYPE_INT>(0);
-    file_field.name = "value";
-    file_field.type = std::make_shared<DataTypeInt32>();
-    std::vector<format::ColumnDefinition> file_schema {file_field};
-
-    auto status = mapper.create_mapping(projected_columns, {}, file_schema);
-    ASSERT_TRUE(status.ok()) << status;
-
-    auto predicate = std::make_shared<Int64BinaryPredicateExpr>(TExprOpcode::GT);
-    predicate->add_child(VSlotRef::create_shared(0, 0, -1, table_column.type, "value"));
-    predicate->add_child(
-            VLiteral::create_shared(table_column.type, Field::create_field<TYPE_STRING>("bad")));
-    format::TableFilter table_filter;
-    table_filter.conjunct = VExprContext::create_shared(predicate);
-    table_filter.global_indices = {format::GlobalIndex(0)};
-
-    format::FileScanRequest file_request;
-    ASSERT_TRUE(
-            mapper.create_scan_request({table_filter}, {}, projected_columns, &file_request, &state)
-                    .ok());
-    ASSERT_EQ(file_request.conjuncts.size(), 1);
-    const auto& localized_expr = file_request.conjuncts[0]->root();
-    ASSERT_EQ(localized_expr->get_num_children(), 2);
-    const auto& localized_child = localized_expr->children()[0];
-    ASSERT_NE(dynamic_cast<const Cast*>(localized_child.get()), nullptr);
-    ASSERT_EQ(localized_child->get_num_children(), 1);
-    const auto* localized_slot = assert_cast<const VSlotRef*>(localized_child->children()[0].get());
-    EXPECT_EQ(localized_slot->column_id(), 0);
-    EXPECT_TRUE(localized_slot->data_type()->equals(*file_field.type));
-    EXPECT_TRUE(localized_child->data_type()->equals(*table_column.type));
-    EXPECT_TRUE(localized_expr->children()[1]->is_literal());
-    EXPECT_TRUE(localized_expr->children()[1]->data_type()->equals(*table_column.type));
-}
-
-TEST_F(CastTest, ColumnMapperDoesNotLeakRewrittenLiteralAcrossSplits) {
-    format::ColumnDefinition table_column;
-    table_column.identifier = Field::create_field<TYPE_INT>(7);
-    table_column.name = "value";
-    table_column.type = std::make_shared<DataTypeInt64>();
-    std::vector<format::ColumnDefinition> projected_columns {table_column};
-
-    auto predicate = std::make_shared<Int64BinaryPredicateExpr>(TExprOpcode::GT);
-    predicate->add_child(VSlotRef::create_shared(0, 0, -1, table_column.type, "value"));
-    predicate->add_child(
-            VLiteral::create_shared(table_column.type, Field::create_field<TYPE_BIGINT>(15)));
-    format::TableFilter table_filter;
-    table_filter.conjunct = VExprContext::create_shared(predicate);
-    table_filter.global_indices = {format::GlobalIndex(0)};
-
-    format::ColumnDefinition int_file_field;
-    int_file_field.identifier = Field::create_field<TYPE_INT>(0);
-    int_file_field.name = "value";
-    int_file_field.type = std::make_shared<DataTypeInt32>();
-    format::TableColumnMapper int_mapper;
-    ASSERT_TRUE(int_mapper.create_mapping(projected_columns, {}, {int_file_field}).ok());
-    format::FileScanRequest int_request;
-    ASSERT_TRUE(int_mapper
-                        .create_scan_request({table_filter}, {}, projected_columns, &int_request,
-                                             &state)
-                        .ok());
-    ASSERT_EQ(int_request.conjuncts.size(), 1);
-    const auto& int_localized_expr = int_request.conjuncts[0]->root();
-    ASSERT_EQ(int_localized_expr->get_num_children(), 2);
-    EXPECT_TRUE(int_localized_expr->children()[1]->is_literal());
-    EXPECT_TRUE(int_localized_expr->children()[1]->data_type()->equals(*int_file_field.type));
-
-    format::ColumnDefinition bigint_file_field;
-    bigint_file_field.identifier = Field::create_field<TYPE_INT>(0);
-    bigint_file_field.name = "value";
-    bigint_file_field.type = std::make_shared<DataTypeInt64>();
-    format::TableColumnMapper bigint_mapper;
-    ASSERT_TRUE(bigint_mapper.create_mapping(projected_columns, {}, {bigint_file_field}).ok());
-    format::FileScanRequest bigint_request;
-    ASSERT_TRUE(bigint_mapper
-                        .create_scan_request({table_filter}, {}, projected_columns, &bigint_request,
-                                             &state)
-                        .ok());
-    ASSERT_EQ(bigint_request.conjuncts.size(), 1);
-    const auto& bigint_localized_expr = bigint_request.conjuncts[0]->root();
-    ASSERT_EQ(bigint_localized_expr->get_num_children(), 2);
-    const auto* localized_slot =
-            assert_cast<const VSlotRef*>(bigint_localized_expr->children()[0].get());
-    EXPECT_EQ(localized_slot->column_id(), 0);
-    EXPECT_TRUE(localized_slot->data_type()->equals(*bigint_file_field.type));
-    EXPECT_TRUE(bigint_localized_expr->children()[1]->is_literal());
-    EXPECT_TRUE(bigint_localized_expr->children()[1]->data_type()->equals(*bigint_file_field.type));
-}
-
-TEST_F(CastTest, ColumnMapperKeepsExplicitSlotCastInSlotLiteralPredicate) {
-    format::TableColumnMapper mapper;
-    format::ColumnDefinition table_column;
-    table_column.identifier = Field::create_field<TYPE_INT>(7);
-    table_column.name = "value";
-    table_column.type = std::make_shared<DataTypeInt64>();
-    std::vector<format::ColumnDefinition> projected_columns {table_column};
-
-    format::ColumnDefinition file_field;
-    file_field.identifier = Field::create_field<TYPE_INT>(0);
-    file_field.name = "value";
-    file_field.type = std::make_shared<DataTypeInt32>();
-    std::vector<format::ColumnDefinition> file_schema {file_field};
-
-    auto status = mapper.create_mapping(projected_columns, {}, file_schema);
-    ASSERT_TRUE(status.ok()) << status;
-
-    auto explicit_cast = Cast::create_shared(std::make_shared<DataTypeString>());
-    explicit_cast->add_child(VSlotRef::create_shared(0, 0, -1, table_column.type, "value"));
-    auto predicate = std::make_shared<Int64BinaryPredicateExpr>(TExprOpcode::GT);
-    predicate->add_child(explicit_cast);
-    predicate->add_child(
-            VLiteral::create_shared(table_column.type, Field::create_field<TYPE_BIGINT>(15)));
-    format::TableFilter table_filter;
-    table_filter.conjunct = VExprContext::create_shared(predicate);
-    table_filter.global_indices = {format::GlobalIndex(0)};
-
-    format::FileScanRequest file_request;
-    ASSERT_TRUE(
-            mapper.create_scan_request({table_filter}, {}, projected_columns, &file_request, &state)
-                    .ok());
-    ASSERT_EQ(file_request.conjuncts.size(), 1);
-    const auto& localized_expr = file_request.conjuncts[0]->root();
-    ASSERT_EQ(localized_expr->get_num_children(), 2);
-    const auto& localized_cast = localized_expr->children()[0];
-    ASSERT_NE(dynamic_cast<const Cast*>(localized_cast.get()), nullptr);
-    EXPECT_TRUE(localized_cast->data_type()->equals(DataTypeString()));
-    ASSERT_EQ(localized_cast->get_num_children(), 1);
-    ASSERT_NE(dynamic_cast<const Cast*>(localized_cast->children()[0].get()), nullptr);
-    const auto* localized_slot =
-            assert_cast<const VSlotRef*>(localized_cast->children()[0]->children()[0].get());
-    EXPECT_EQ(localized_slot->column_id(), 0);
-    EXPECT_TRUE(localized_slot->data_type()->equals(*file_field.type));
-}
-
-TEST_F(CastTest, ColumnMapperDoesNotNestCastFilterAcrossScanRequests) {
-    format::TableColumnMapper mapper;
-    format::ColumnDefinition table_column;
-    table_column.identifier = Field::create_field<TYPE_INT>(7);
-    table_column.name = "value";
-    table_column.type = std::make_shared<DataTypeInt64>();
-    std::vector<format::ColumnDefinition> projected_columns {table_column};
-
-    format::ColumnDefinition file_field;
-    file_field.identifier = Field::create_field<TYPE_INT>(0);
-    file_field.name = "value";
-    file_field.type = std::make_shared<DataTypeInt32>();
-    std::vector<format::ColumnDefinition> file_schema {file_field};
-
-    auto status = mapper.create_mapping(projected_columns, {}, file_schema);
-    ASSERT_TRUE(status.ok()) << status;
-
-    auto predicate = std::make_shared<Int64ChildGreaterThanExpr>(15);
-    predicate->add_child(VSlotRef::create_shared(0, 0, -1, table_column.type, "value"));
-    format::TableFilter table_filter;
-    table_filter.conjunct = VExprContext::create_shared(predicate);
-    table_filter.global_indices = {format::GlobalIndex(0)};
-
-    format::FileScanRequest first_request;
-    ASSERT_TRUE(mapper.create_scan_request({table_filter}, {}, projected_columns, &first_request,
-                                           &state)
-                        .ok());
-    format::FileScanRequest second_request;
-    ASSERT_TRUE(mapper.create_scan_request({table_filter}, {}, projected_columns, &second_request,
-                                           &state)
-                        .ok());
-
-    ASSERT_EQ(second_request.conjuncts.size(), 1);
-    const auto& localized_expr = second_request.conjuncts[0]->root();
-    ASSERT_EQ(localized_expr->get_num_children(), 1);
-    const auto& localized_child = localized_expr->children()[0];
-    ASSERT_NE(dynamic_cast<const Cast*>(localized_child.get()), nullptr);
-    ASSERT_EQ(localized_child->get_num_children(), 1);
-    const auto* localized_slot = assert_cast<const VSlotRef*>(localized_child->children()[0].get());
-    EXPECT_EQ(localized_slot->column_id(), 0);
-}
-
-TEST_F(CastTest, ColumnMapperRewritesPreviousCastFilterToMatchingSplitType) {
-    format::ColumnDefinition table_column;
-    table_column.identifier = Field::create_field<TYPE_INT>(7);
-    table_column.name = "value";
-    table_column.type = std::make_shared<DataTypeInt64>();
-    std::vector<format::ColumnDefinition> projected_columns {table_column};
-
-    auto predicate = std::make_shared<Int64ChildGreaterThanExpr>(15);
-    predicate->add_child(VSlotRef::create_shared(0, 0, -1, table_column.type, "value"));
-    format::TableFilter table_filter;
-    table_filter.conjunct = VExprContext::create_shared(predicate);
-    table_filter.global_indices = {format::GlobalIndex(0)};
-
-    format::ColumnDefinition int_file_field;
-    int_file_field.identifier = Field::create_field<TYPE_INT>(0);
-    int_file_field.name = "value";
-    int_file_field.type = std::make_shared<DataTypeInt32>();
-
-    format::TableColumnMapper int_mapper;
-    ASSERT_TRUE(int_mapper.create_mapping(projected_columns, {}, {int_file_field}).ok());
-    format::FileScanRequest int_request;
-    ASSERT_TRUE(int_mapper
-                        .create_scan_request({table_filter}, {}, projected_columns, &int_request,
-                                             &state)
-                        .ok());
-
-    const auto& int_localized_expr = int_request.conjuncts[0]->root();
-    ASSERT_EQ(int_localized_expr->get_num_children(), 1);
-    ASSERT_NE(dynamic_cast<const Cast*>(int_localized_expr->children()[0].get()), nullptr);
-
-    format::ColumnDefinition bigint_file_field;
-    bigint_file_field.identifier = Field::create_field<TYPE_INT>(0);
-    bigint_file_field.name = "value";
-    bigint_file_field.type = std::make_shared<DataTypeInt64>();
-
-    format::TableColumnMapper bigint_mapper;
-    ASSERT_TRUE(bigint_mapper.create_mapping(projected_columns, {}, {bigint_file_field}).ok());
-    format::FileScanRequest bigint_request;
-    ASSERT_TRUE(bigint_mapper
-                        .create_scan_request({table_filter}, {}, projected_columns, &bigint_request,
-                                             &state)
-                        .ok());
-
-    const auto& bigint_localized_expr = bigint_request.conjuncts[0]->root();
-    ASSERT_EQ(bigint_localized_expr->get_num_children(), 1);
-    const auto& bigint_localized_child = bigint_localized_expr->children()[0];
-    const auto* localized_slot = assert_cast<const VSlotRef*>(bigint_localized_child.get());
-    EXPECT_EQ(localized_slot->column_id(), 0);
-    EXPECT_TRUE(localized_slot->data_type()->equals(*bigint_file_field.type));
-
-    Block block;
-    block.insert(ColumnHelper::create_column_with_name<DataTypeInt64>({11, 22}));
-    auto* conjunct = bigint_request.conjuncts[0].get();
-    auto status = conjunct->prepare(&state, RowDescriptor());
-    ASSERT_TRUE(status.ok()) << status;
-    status = conjunct->open(&state);
-    ASSERT_TRUE(status.ok()) << status;
-    IColumn::Filter filter(block.rows(), 1);
-    bool can_filter_all = false;
-    status = conjunct->execute_filter(&block, filter.data(), block.rows(), false, &can_filter_all);
-    ASSERT_TRUE(status.ok()) << status;
-    EXPECT_FALSE(can_filter_all);
-    ASSERT_EQ(filter.size(), 2);
-    EXPECT_EQ(filter[0], 0);
-    EXPECT_EQ(filter[1], 1);
-    conjunct->close();
-}
-
-TEST_F(CastTest, ColumnMapperKeepsTableSlotIdWhenFileBlockPositionChanges) {
-    format::ColumnDefinition table_column;
-    table_column.identifier = Field::create_field<TYPE_INT>(7);
-    table_column.name = "value";
-    table_column.type = std::make_shared<DataTypeInt64>();
-    std::vector<format::ColumnDefinition> projected_columns {table_column};
-
-    format::ColumnDefinition file_field;
-    file_field.identifier = Field::create_field<TYPE_INT>(10);
-    file_field.name = "value";
-    file_field.type = std::make_shared<DataTypeInt64>();
-
-    format::TableColumnMapper mapper;
-    ASSERT_TRUE(mapper.create_mapping(projected_columns, {}, {file_field}).ok());
-
-    auto predicate = std::make_shared<Int64ChildGreaterThanExpr>(15);
-    predicate->add_child(VSlotRef::create_shared(0, 0, -1, table_column.type, "value"));
-    format::TableFilter table_filter;
-    table_filter.conjunct = VExprContext::create_shared(predicate);
-    table_filter.global_indices = {format::GlobalIndex(0)};
-
-    format::FileScanRequest first_request;
-    ASSERT_TRUE(mapper.localize_filters({table_filter}, {}, &first_request, &state).ok());
-    ASSERT_EQ(first_request.conjuncts.size(), 1);
-    const auto* first_slot =
-            assert_cast<const VSlotRef*>(first_request.conjuncts[0]->root()->children()[0].get());
-    EXPECT_EQ(first_slot->slot_id(), 7);
-    EXPECT_EQ(first_slot->column_id(), 0);
-
-    format::FileScanRequest second_request;
-    second_request.local_positions.emplace(format::LocalColumnId(9), format::LocalIndex(0));
-    second_request.local_positions.emplace(format::LocalColumnId(10), format::LocalIndex(1));
-    second_request.non_predicate_columns.push_back(field_projection(9));
-    ASSERT_TRUE(mapper.localize_filters({table_filter}, {}, &second_request, &state).ok());
-    ASSERT_EQ(second_request.conjuncts.size(), 1);
-    const auto* second_slot =
-            assert_cast<const VSlotRef*>(second_request.conjuncts[0]->root()->children()[0].get());
-    EXPECT_EQ(second_slot->slot_id(), 7);
-    EXPECT_EQ(second_slot->column_id(), 1);
-
-    Block block;
-    block.insert(ColumnHelper::create_column_with_name<DataTypeInt64>({100, 100}));
-    block.insert(ColumnHelper::create_column_with_name<DataTypeInt64>({11, 22}));
-    auto* conjunct = second_request.conjuncts[0].get();
-    auto status = conjunct->prepare(&state, RowDescriptor());
-    ASSERT_TRUE(status.ok()) << status;
-    status = conjunct->open(&state);
-    ASSERT_TRUE(status.ok()) << status;
-    IColumn::Filter filter(block.rows(), 1);
-    bool can_filter_all = false;
-    status = conjunct->execute_filter(&block, filter.data(), block.rows(), false, &can_filter_all);
-    ASSERT_TRUE(status.ok()) << status;
-    EXPECT_FALSE(can_filter_all);
-    ASSERT_EQ(filter.size(), 2);
-    EXPECT_EQ(filter[0], 0);
-    EXPECT_EQ(filter[1], 1);
-    conjunct->close();
 }
 
 } // namespace doris::format
